requirements.txt
django==4.1.5
gunicorn==20.1.0
mysqlclient
python-dotenv
google-cloud-storage
google-cloud-secret-manager
django-storages